projChoice user auth added

+++ server/routes/projChoices.js
@@ -1,9 +1,10 @@
 const sequenceGenerator = require('./sequenceGenerator');
 const ProjChoice = require('../models/projChoice');
+const checkAuth = require("../middleware/check-auth")
+
 
 var express = require('express');
 var router = express.Router();
-module.exports = router;
 
 router.get("/", (req, res, next) => {
   ProjChoice.find()
@@ -14,10 +15,15 @@ router.get("/", (req, res, next) => {
       })
     }
     return res.status(200).json(projChoices);
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching project choices failed"
+    });
   });
 });
 
-router.post('/', (req, res, next) => {
+router.post('/', checkAuth, (req, res, next) => {
   console.log('projChoices posted?');
   const maxProjChoiceId = sequenceGenerator.nextId("projChoices");
 
@@ -25,7 +31,8 @@ router.post('/', (req, res, next) => {
     id: maxProjChoiceId,
     name: req.body.name,
     proposedBy: req.body.proposedBy,
-    status: req.body.status
+    status: req.body.status,
+    creator: req.userData.userId
   });
   console.log("projChoice, look for body status", projChoice);
 
@@ -44,18 +51,23 @@ router.post('/', (req, res, next) => {
     });
 });
 
-router.put('/:id', (req, res, next) => {
+router.put('/:id', checkAuth, (req, res, next) => {
   ProjChoice.findOne({ id: req.params.id })
     .then(projChoice => {
       projChoice.name = req.body.name;
       projChoice.proposedBy = req.body.proposedBy;
       projChoice.status = req.body.status;
+      projChoice.creator = req.userData.userId;
 
-      ProjChoice.updateOne({ id: req.params.id }, projChoice)
+      ProjChoice.updateOne({ id: req.params.id , creator: req.userData.userId}, projChoice)
         .then(result => {
-          res.status(204).json({
-            message: 'ProjChoice updated successfully'
-          })
+          console.log(result);
+          if(result.modifiedCount > 0){
+            res.status(204).json({ message: "ProjChoice updated successfully!"});
+            console.log("the user updated their project!");
+          } else {
+            res.status(401).json({ message: "Not authorized!"});
+          }
         })
         .catch(error => {
            res.status(500).json({
@@ -72,10 +84,10 @@ router.put('/:id', (req, res, next) => {
     });
 });
 
-router.delete("/:id", (req, res, next) => {
+router.delete("/:id", checkAuth, (req, res, next) => {
   ProjChoice.findOne({ id: req.params.id })
     .then(projChoice => {
-      ProjChoice.deleteOne({ id: req.params.id })
+      ProjChoice.deleteOne({ id: req.params.id, creator: req.userData.userId })
         .then(result => {
           res.status(204).json({
             message: "ProjChoice deleted successfully"
@@ -83,7 +95,7 @@ router.delete("/:id", (req, res, next) => {
         })
         .catch(error => {
            res.status(500).json({
-           message: 'An error occurred',
+           message: 'Deletion failed',
            error: error
          });
         })
@@ -95,3 +107,5 @@ router.delete("/:id", (req, res, next) => {
       });
     });
 });
+
+module.exports = router;
